refactor: refactor cart, order, and reviews module

+++ src/modules/products/products.controller.ts
@@ -1,8 +1,8 @@
 import { CartEntity, CartItemsEntity, CategoryProductEntity, OrderEntity, OrderItemEntity, ProductEntity, ReviewsEntity } from "@entities";
 import { InitRepository, InjectRepositories } from "@helpers";
 import { TRequest, TResponse } from "@types";
-import { Between, In, LessThanOrEqual, Like, MoreThanOrEqual, Repository } from "typeorm";
-import { CategoryDto, ProductDto, ReviewsDto } from "./dto";
+import { Between, In, Like, MoreThanOrEqual, Repository } from "typeorm";
+import { CategoryDto, ProductDto } from "./dto";
 
 interface CustomError extends Error {
   statusCode?: number;
@@ -12,23 +12,8 @@ export class ProductController {
   @InitRepository(ProductEntity)
   productRepository: Repository<ProductEntity>;
 
-  @InitRepository(ReviewsEntity)
-  reviewsRepository: Repository<ReviewsEntity>;
-
-  @InitRepository(OrderItemEntity)
-  orderItemRepository: Repository<OrderItemEntity>;
-
-  @InitRepository(OrderEntity)
-  orderRepository: Repository<OrderEntity>;
-
   @InitRepository(CategoryProductEntity)
   categoryProductRepository: Repository<CategoryProductEntity>;
-
-  @InitRepository(CartItemsEntity)
-  cartitemsRepository: Repository<CartItemsEntity>;
-
-  @InitRepository(CartEntity)
-  cartRepository: Repository<CartEntity>;
 
   constructor() {
     InjectRepositories(this);
@@ -44,7 +29,7 @@ export class ProductController {
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 
@@ -60,7 +45,7 @@ export class ProductController {
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 
@@ -105,14 +90,19 @@ export class ProductController {
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 
   //delete existed product
   public deleteProduct = async (req: TRequest, res: TResponse): Promise<void> => {
     try {
-      const prodId = req.params.productId;
+      const prodId = Number(req.params.productId);
+      const product = await this.productRepository.findOne({ where: { id: prodId } });
+      if (!product) {
+        res.status(404).json({ message: "No Product Found" });
+        return;
+      }
       await this.productRepository.delete(prodId);
       res.status(200).json({ success: true, message: "product deleted successfully" });
     } catch (err: unknown) {
@@ -120,7 +110,7 @@ export class ProductController {
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 
@@ -135,8 +125,13 @@ export class ProductController {
 
   //add the categories
   public addCategory = async (req: TRequest<CategoryDto>, res: TResponse) => {
-    const { name, description } = req.dto;
+    const { name } = req.dto;
     try {
+      const isExistcategory = await this.categoryProductRepository.findOne({ where: { name: name } });
+      if (isExistcategory) {
+        res.status(400).json({ message: "Category already exist" });
+        return;
+      }
       const category = await this.categoryProductRepository.create(req.dto);
       this.categoryProductRepository.save(category);
       res.status(201).json({ success: true, category: category });
@@ -145,7 +140,7 @@ export class ProductController {
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 
@@ -191,10 +186,11 @@ export class ProductController {
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 
+  //fetch all the trending products
   public trendingProducts = async (req: TRequest, res: TResponse): Promise<void> => {
     try {
       const trendProducts = await this.productRepository.find({
@@ -211,61 +207,10 @@ export class ProductController {
       res.status(200).json({ success: true, trendProducts });
     } catch (err: unknown) {
       const error = err as CustomError;
-      error.statusCode = error.statusCode || 500;
-      res.status(400).json({ error });
-    }
-  };
-
-  //review posting
-  public postReview = async (req: TRequest<ReviewsDto>, res: TResponse) => {
-    try {
-      const productId = Number(req.params.productId);
-      const prod = await this.productRepository.findOne({ where: { id: productId } });
-      if (!prod) {
-        const error = new Error("No product") as CustomError;
-        error.statusCode = 404;
-        throw error;
-      }
-      const { rating, review } = req.dto;
-      const reviews = await this.reviewsRepository.create({
-        rating,
-        review,
-        userId: req.user.id,
-        productId,
-      });
-      this.reviewsRepository.save(reviews);
-      res.status(200).json({ success: true, reviews });
-    } catch (err: unknown) {
-      const error = err as CustomError;
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
-    }
-  };
-
-  //getting review
-  public getReview = async (req: TRequest, res: TResponse): Promise<void> => {
-    try {
-      const productId = Number(req.params.productId);
-      const prod = await this.productRepository.find({ where: { id: productId } });
-      if (!prod) {
-        const error = new Error("No product") as CustomError;
-        error.statusCode = 404;
-        throw error;
-      }
-      const review = await this.reviewsRepository.findOne({ where: { productId: productId } });
-      if (!review) {
-        res.status(404).json({ message: "no reviews" });
-        return;
-      }
-      res.status(200).json({ success: true, review });
-    } catch (err: unknown) {
-      const error = err as CustomError;
-      if (!error.statusCode) {
-        error.statusCode = 500;
-      }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 
@@ -284,250 +229,10 @@ export class ProductController {
       res.status(200).json({ success: true, product, category });
     } catch (err: unknown) {
       const error = err as CustomError;
-      error.statusCode = error.statusCode || 500;
-      res.status(400).json({ error });
-    }
-  };
-
-  // post Cart
-  public postCart = async (req: TRequest, res: TResponse): Promise<void> => {
-    try {
-      const { productId, quantity } = req.body;
-      if (!productId || !quantity) {
-        const error = new Error("Product and Quantity required") as CustomError;
-        error.statusCode = 400;
-        throw error;
-      }
-
-      // Check if the user has an active cart
-      let cart = await this.cartRepository.findOne({ where: { user: { id: req.user.id } } });
-
-      if (!cart) {
-        cart = this.cartRepository.create({ user: { id: req.user.id } });
-        await this.cartRepository.save(cart);
-      }
-
-      // Check if the product is already in the cart
-      let cartItem = await this.cartitemsRepository.findOne({
-        where: { cart: { id: cart.id }, product: { id: productId } },
-      });
-
-      if (cartItem) {
-        cartItem.quantity += quantity;
-      } else {
-        cartItem = this.cartitemsRepository.create({
-          quantity: quantity,
-          productId: productId,
-          cartId: cart.id,
-        });
-      }
-
-      await this.cartitemsRepository.save(cartItem);
-
-      res.status(200).json({ success: true, cartItem, message: "Product added to cart" });
-    } catch (err: unknown) {
-      const error = err as CustomError;
-      error.statusCode = error.statusCode || 500;
-      res.status(400).json({ error });
-    }
-  };
-
-  // get Cart
-  public getCart = async (req: TRequest, res: TResponse): Promise<void> => {
-    try {
-      // Find the cart for the current user
-      const cart = await this.cartRepository.findOne({
-        where: { userId: req.user.id },
-      });
-      if (!cart || cart === null) {
-        res.status(404).json({ message: "No Product Found" });
-        return;
-      }
-
-      // Find all cart items associated with the cart
-      const cartItems = await this.cartitemsRepository.find({
-        where: { cart: { id: cart.id } },
-        relations: ["product"],
-        select: {
-          id: true,
-          quantity: true,
-          product: { id: true, name: true, price: true, rating: true, discount: true },
-        },
-      });
-
-      if (!cartItems.length) {
-        const error = new Error("No Products in Cart") as CustomError;
-        error.statusCode = 404;
-        throw error;
-      }
-
-      res.status(200).json({ success: true, cart, cartItems });
-    } catch (err: unknown) {
-      const error = err as CustomError;
-      error.statusCode = error.statusCode || 500;
-      res.status(400).json({ error });
-    }
-  };
-
-  // create order
-  public postOrder = async (req: TRequest, res: TResponse) => {
-    try {
-      // Find the cart for the current user
-      const cart = await this.cartRepository.findOne({
-        where: { userId: req.user.id },
-      });
-      if (!cart) {
-        const error = new Error("No Cart Found") as CustomError;
-        error.statusCode = 404;
-        throw error;
-      }
-      const cartItems = await this.cartitemsRepository.find({
-        where: { cart: { id: cart.id } },
-        relations: ["product"],
-      });
-
-      if (!cartItems || cartItems.length === 0) {
-        const error = new Error("No Products in Cart") as CustomError;
-        error.statusCode = 400;
-        throw error;
-      }
-      const order = await this.orderRepository.create({
-        userId: req.user.id,
-        totalAmount: cartItems.reduce((acc: number, item: any) => {
-          return acc + item.quantity * item.product.price;
-        }, 0),
-
-        isCancelled: false,
-        status: "Pending",
-      });
-      await this.orderRepository.save(order);
-      const orderItems = await cartItems.map((item: any) => ({
-        orderId: order.id,
-        productId: item.productId,
-        quantity: item.quantity,
-        price: item.product.price,
-      }));
-      await this.orderItemRepository.insert(orderItems);
-      await this.cartitemsRepository.remove(cartItems);
-      res.status(200).json({ success: true, order: order });
-    } catch (err: unknown) {
-      const error = err as CustomError;
       if (!error.statusCode) {
         error.statusCode = 500;
       }
-      res.status(400).json({ error: error });
-    }
-  };
-
-  //remove item from cart
-  public removeCartItems = async (req: TRequest, res: TResponse): Promise<void> => {
-    try {
-      const userId = req.user.id;
-      const productId = req.body.productId;
-
-      // Find the cart associated with the user
-      const cart = await this.cartRepository.findOne({ where: { userId } });
-      if (!cart) {
-        throw new Error("Cart not found") as CustomError;
-      }
-
-      // Find the cart item within the user's cart
-      const existProduct = await this.cartitemsRepository.findOne({
-        where: { cart: { id: cart.id }, product: { id: productId } },
-      });
-
-      if (!existProduct) {
-        const error = new Error("There is no such item in the cart") as CustomError;
-        error.statusCode = 404;
-        throw error;
-      }
-
-      // Remove the cart item
-      await this.cartitemsRepository.remove(existProduct);
-
-      res.status(200).json({ success: true, message: "Item removed" });
-    } catch (err: unknown) {
-      const error = err as CustomError;
-      error.statusCode = error.statusCode || 500;
-      res.status(400).json({ error });
-    }
-  };
-
-  // getOrders
-  public getOrder = async (req: TRequest, res: TResponse): Promise<void> => {
-    try {
-      const orders = await this.orderRepository.find({
-        where: { userId: req.user.id },
-        select: ["id", "isCancelled", "status"],
-      });
-
-      if (!orders.length) {
-        res.status(404).json({ message: "You haven't ordered yet!" });
-        return;
-      }
-
-      const orderIds = orders.map(order => order.id);
-
-      const orderItems = await this.orderItemRepository.find({
-        where: { order: { id: In(orderIds) } },
-        select: ["id", "orderId", "productId", "quantity"],
-        relations: ["product"],
-      });
-
-      const formattedOrders = orders.map(order => ({
-        id: order.id,
-        isCancelled: order.isCancelled,
-        status: order.status,
-        orderItems: orderItems
-          .filter(item => item.orderId === order.id)
-          .map(item => ({
-            productId: item.product?.id,
-            quantity: item.quantity,
-            price: item.product?.price,
-          })),
-      }));
-
-      res.status(200).json({ success: true, orders: formattedOrders });
-    } catch (err: unknown) {
-      const error = err as CustomError;
-      error.statusCode = error.statusCode || 500;
-      res.status(400).json({ error });
-    }
-  };
-
-  public orderDetails = async (req: TRequest, res: TResponse): Promise<void> => {
-    try {
-      const orderId = Number(req.params.orderId);
-      const action = req.query.action;
-
-      if (action === "cancel") {
-        await this.orderRepository.update(orderId, { isCancelled: true, status: "Cancelled" });
-      }
-
-      const order = await this.orderRepository.findOne({ where: { id: orderId } });
-      if (!order) {
-        const error = new Error("No orders yet") as CustomError;
-        error.statusCode = 404;
-        throw error;
-      }
-      const orderDetails = await this.orderItemRepository.find({
-        where: { order: { id: orderId } },
-        select: ["productId", "quantity"],
-        relations: ["product"],
-      });
-
-      const orderStatus = await this.orderRepository.findOne({
-        where: { id: orderId },
-        select: ["status"],
-      });
-
-      res.status(200).json({ success: true, orderDetails, orderStatus });
-    } catch (err: unknown) {
-      const error = err as CustomError;
-      if (!error.statusCode) {
-        error.statusCode = 500;
-      }
-      res.status(400).json({ error: error });
+      res.status(error.statusCode).json({ error: error });
     }
   };
 }
